Introduce SourceMethodSymbolWithAttributes
